Add a Thread::Backtrace::Location#pretty_name method
This commit adds a new method #pretty_name to the
Thread::Backtrace::Location class. This method returns a string which
contains not only the method name of the frame, but also the class name
on which it is defined. This helps understand the context of a backtrace
without nescessarily needing to look back to the code.

The implementation of this uses similar formatting to
rb_profile_frame_qualified_method_name.