Handle pretty backtraces when in class/module definition
When a backtrace is obtained from a "class Foo; ...; end;" (or module)
block, we print "class exec in Foo" or "module exec in Foo::Bar".

This can't really be implemented properly for a couple of reasons:
* There's nothing on the iseq that points to the actual class/module; at
  compile time all we know is the name (from the "class Foo" line). The
  class/module is the receiver when the thing is actually executed.
  (Although - we _could_ get the receiver from the call frame).
* The iseq only has access to the lexical context of the class
  hierarchy, not the actual namespace nesting. That might be different
  if e.g. we're inside a "class_eval" or something.

I hacked it as best as I could by parsing out the location label.