Make the "pretty method name" output the default
This is implemented by changing the implementation of
Thread::Backtrace::Location#inspect to use the pretty_name string in
its output. The existing components should remain on the Location
structure in their original form; the only applications that should be
affected by this change are those that are parsing the string output
format of backtraces instead of using Backtrace::Location.