Add _ext structure to rb_method_entry_t to support addiitonal fields
rb_method_entry_t (and its const cousin rb_callable_method_entry_t) need
to be RVALUEs (they are stored in T_IMEMO objects managed by the GC).
However, these structs are currently using all five words of the RVALUE.
Thus, it is not possible to add additional fields to them.

In order to solve this, we define a rb_method_entry_ext_t structure to
hold additional attributes and manage it in a similar way to how
rb_classext_struct is managed for RClass.

If USE_RVARGC is on (i.e. multiple size pools in the GC are enabled),
then we store the method entry in a larger size pool and place the
rb_method_entry_ext_t data inline with the object. Otherwise, we store a
pointer to a C heap-allocated rb_method_entry_ext_t in one of the RVALUE
words.

The details of either method are abstracted behind  METHOD_ENTRY_EXT()
and CALLABLE_METHOD_ENTRY_EXT() macros, which act analogously to the
RCLASS_EXT() one.

In order to make room for the ext pointer in non-RVARGC configurations,
the "owner" field of rb_method_entry_t has been oved to the
rb_method_entry_ext_t structure.