Move __{tmp_}classpath__ attrs to rb_classext_t
This commit removes the hidden __classpath__ and __tmp_classpath__
instance variables stored in class/module objects. They are moved into a
single classpath string stored on the rb_classext_t structure, as well
as a flag bit to singify whether this is a "permanent" classpath or a
"temporary" one.

All classes/modules assigned to a constant already have these set, so
moving them shouldn't result in a net increase in memory usage there.

Unfortunately these ivars were _not_ being set on T_ICLASS objects, so
there may be a small increase in memory usage from included modules by
virture of having to carry this VALUE around in rb_classext_t.

The motivation for this work is to allow fast lookups of class names in
very hot codepaths and exposing them this to profiling tools.