Implement a Socket::Credentials struct for handling unix socket creds
Depending on the platform, Unix sockets can support fetching the
credentials of a remote peer, or pushing credentials to a remote peer,
through a variety of socket options and ancillary message types.

Currently, Ruby is able to parse these ancillary data and socket option
structures, but the only way this is exposed back to Ruby code is
through the #inspect method. Parsing inspect strings is not a robust way
to write programs!

This change adds a type Socket::Credentials, which is a structure for
holding the various kinds of credentials that unixen support passing
between processes. It also implements Socket::AncillaryData#credentials
and Socket::Option#credentials, to allow parsing these credentials from
ancillary data/socket options respectively.