Fix thread unsafety with rb_register_postponed_job
rb_register_postponed_job is designed to be both threadsafe and
async-signal-safe. That is to say, it should be legal to call it in all
of the following circumstances:

* From a thread holding the GVL
* From a Ruby thread not holding the GVL
* From a native thread not otherwise known to the Ruby VM
* From a signal handler, interrupting non-ruby code
* From a signal handler, interrupting the Ruby VM
* From a signal handler, interrupting an execution of
  rb_register_postponed_job (i.e. it must be _reentrant_)
* From a signal handler, interrupting rb_postponed_job_flush

The current implementation is unfortunately racy when called from a
thread whilst rb_postponed_job_flush is running in another thread. It is
possible for one thread executing postponed jobs to decrement
vm->postponed_job_index, and then for another thread to re-incremenet it
& overwrite the relevant slot in vm->postponed_job_buffer before the
first thread goes to execute the original job. This can lead to a
function being called with an argument intended for a _different_
function.

This commit fixes the issue by re-implementing both the postponed-job
and workqueue mechanism, and wrapping them into an overarching
`rb_postponed_job_queues_t` structure. The structure is split into
"blocking" and "async" parts.

The "blocking" component is a re-implementation of the workqueue
mechanism. Instead of using a linked-list of jobs, it stores them in a
hashtable, so that fast deduplication of jobs in _one functions is
possible. This is an approach which Eric Wong originally committed in
5a1dfb04, but had to be reverted due to async-signal-safety concerns.
When protected by a mutex, however, this is perfectly fine.

The "blocking" queue is exposed through rb_workqueue_register and
rb_workqueue_register_one; these functions are safe to be called at any
time and from any thread, but _NOT_ from a signal handler (since they
use a mutex). The job is guaranteed to be enqueued; the function may block
or allocate memory in order to fulfil that guarantee.

The "async" component is a re-implementation of the existing postponed
job buffer. It represents a lock-free, single-consumer, multi-producer
ringbuffer. It's safe to insert jobs into this ringbuffer at any time,
from any thread, even in a signal handler. Because allocating memory is
not safe inside a signal handler, the ringbuffer is fixed-size and it's
possible enqueuing jobs might fail for want of space; callers must
accept that.

The "async" queue is exposed through the rb_postponed_job_register and
rb_postponed_job_register_one functions. These functions can be called
from any thread, including signal handlers.

The _execution_ of both the async and blocking jobs is unified in
rb_postponed_job_flush. The async queue is actually drained into the
blocking queue, which has the effect of deduplicating jobs with the same
callback enqueued with rb_workqueue_register_one and
rb_postponed_job_register_one.

[Bug #19991]