Instrument getaddrinfo calls (both with and without threading)
This adds instrumentation (both in the new, threaded resolver path with
GETADDRINFO_IMPL=2, and the old, single-threaded version with
GETADDRINFO_IMPL=1) to getaddrinfo calls.

To use it, set `Addrinfo.stats_cb=` to a proc, which will receive an
instance of the Addrinfo::Stats struct:

```
Addrinfo.stats_cb = proc do |stats|
    # Do something with stats!
    $statsd.distribution('getaddrinfo.total_time', stats.total)
    # ...
end
```

The stats struct has five fields:
* `#time` - the total time the call took from Ruby's perspective
* `#getaddrinfo3_time` - the time spent inside libc's `getaddrinfo(3)`
  function. This won't include time starting the thread, waiting for the
  GVL afterwards, etc.
* `#start_delay` - the time difference between when Ruby called
  `Addrinfo.getaddrinfo` and when we actually called libc's
  `getaddrinfo(3) function. This will be the time taken to spin up the
   resolver thread and switch to it.
* `#finish_delay` - the time difference between when libc returned from
  getaddrinfo  and when the Ruby function returned. This will include
  the time taken to re-acquire the GVL.
* `#interrupted` - true if the resolver was interrupted by a timeout or
  some other interrupt, otherwise false.

Note that some of the times (other than `#time`) can be zero in cases
where an interrupt happened (depending on the resolver implementation
and when the interrupt happened); make sure to exclude zeros from being
sent to statsd I think.